refactored editevent component to be a page component instead to simplify the logic.
